# Make the current group size a variable updated by reference