--- a/NEWS.md
+++ b/NEWS.md
@@ -1,5 +1,7 @@
 # dplyr (development version)
 
+* `n()` is now a little faster when there are many groups (#6727).
+
 * `if_else()` and `case_when()` again accept logical conditions that have
   attributes (#6678).