--- a/src/init.cpp
+++ b/src/init.cpp
@@ -42,7 +42,8 @@ SEXP get_names_summarise_recycle_chunks(){
 SEXP symbols::ptype = Rf_install("ptype");
 SEXP symbols::levels = Rf_install("levels");
 SEXP symbols::groups = Rf_install("groups");
-SEXP symbols::dot_current_group = Rf_install(".current_group");
+SEXP symbols::current_group_id = Rf_install("dplyr:::current_group_id");
+SEXP symbols::current_group_size = Rf_install("dplyr:::current_group_size");
 SEXP symbols::current_expression = Rf_install("current_expression");
 SEXP symbols::rows = Rf_install("rows");
 SEXP symbols::caller = Rf_install("caller");
@@ -50,13 +51,13 @@ SEXP symbols::current_data = Rf_install("current_data");
 SEXP symbols::dot_drop = Rf_install(".drop");
 SEXP symbols::dplyr_internal_error = Rf_install("dplyr_internal_error");
 SEXP symbols::dplyr_internal_signal = Rf_install("dplyr_internal_signal");
-SEXP symbols::dot_indices = Rf_install(".indices");
 SEXP symbols::chops = Rf_install("chops");
 SEXP symbols::vec_is_list = Rf_install("vec_is_list");
 SEXP symbols::new_env = Rf_install("new.env");
 SEXP symbols::dot_data = Rf_install(".data");
 SEXP symbols::used = Rf_install("used");
 SEXP symbols::across = Rf_install("across");
+SEXP symbols::env_current = Rf_install("env_current");
 SEXP symbols::env_mask_bindings = Rf_install("env_mask_bindings");
 
 SEXP vectors::classes_vctrs_list_of = get_classes_vctrs_list_of();
@@ -114,7 +115,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"dplyr_mask_binding_remove", (DL_FUNC)& dplyr_mask_binding_remove, 2},
   {"dplyr_mask_binding_add", (DL_FUNC)& dplyr_mask_binding_add, 4},
 
-  {"dplyr_lazy_vec_chop_impl", (DL_FUNC)& dplyr_lazy_vec_chop, 4},
+  {"dplyr_lazy_vec_chop_impl", (DL_FUNC)& dplyr_lazy_vec_chop, 5},
   {"dplyr_make_mask_bindings", (DL_FUNC)& dplyr_make_mask_bindings, 2},
   {"env_resolved", (DL_FUNC)& env_resolved, 2},