This PR:

- Provides access to the current group size through `$get_current_group_size()`
- Uses this helper in `n()`, which makes it a little faster

This also gives us an easy private way to access the current group size for https://github.com/tidyverse/dplyr/issues/6685#issuecomment-1430392850

It also gets rid of the need to do `parent.env(chops)` internally, which always felt a little clunky to me.

``` r
library(dplyr)

set.seed(123)

# ~100k groups, so that many evaluations of `n()`
x <- sample(1e5, 1e6, replace = TRUE)
df <- tibble(x = x)
df <- group_by(df, x)
df
#> # A tibble: 1,000,000 × 1
#> # Groups:   x [99,995]
#>        x
#>    <int>
#>  1 51663
#>  2 57870
#>  3  2986
#>  4 29925
#>  5 95246
#>  6 68293
#>  7 62555
#>  8 45404
#>  9 65161
#> 10 46435
#> # … with 999,990 more rows

bench::mark(summarise(df, n = n()), iterations = 100)

# Main
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 summarise(df, n = n())    919ms    1.14s     0.873    7.15MB     13.8

# This PR
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 summarise(df, n = n())    596ms    742ms      1.35    7.16MB     15.1
```

<sup>Created on 2023-02-15 with [reprex v2.0.2.9000](https://reprex.tidyverse.org)</sup>